bhyve/passthru: keep physical and virtual COMMAND reg in sync
On startup the virtual command register is initialized with following
values:
If a device has a IO BAR, IO decoding will be enabled in the virtual
command register.
If a device has a memory BAR, memory decoding will be enabled in the
virtual command register.

This virtual command register can differ from the physical value of
the passthru device. To avoid issues, keep them synchronized.
Write the value of the virtual command register to the physical one.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>