vt: add a vt_dummy driver
By default vt uses vt_efifb if no video driver is loaded. vt_efifb grabs
the EFI framebuffer and writes directly to it. When the graphics driver
is unloaded from the GPU, vt_efifb still writes to the EFI framebuffer.
If GPU passthrough for a bhyve VM is used, the guest uses the EFI
framebuffer too. When vt_efifb writes new data to the EFI framebuffer it
results in stripes on the guest display.  This can be avoided by loading
this new vt_dummy driver that ignores all draw calls.

Note: As vt_dummy ignores all draw call, loading it will freeze the
display.