bhyve/FwCfg: replace bhyve fwctl by qemu fwcfg
QEMU's fwcfg is much more powerfull than bhyve's fwctl. Additionally,
bhyve's fwctl is just used to pass the number of vcpus to the
guest. It could be easily done by QEMU's fwcfg. Sadly, bhyve
decided to use the same ports for fwctl as QEMU's fwcfg. It's not
possible to use both interfaces.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>