bhyve/basl: add qemu_loader
The hypervisor is aware of all system properties. For the guest bios
it's hard and complex to detect all system properties. For that reason,
it would be better if the hypervisor creates ACPI tables instead of the
guest. Therefore, the hypervisor has to send the ACPI tables to the
guest. At the moment, bhyve just copies the ACPI tables into the guest
memory. This approach has much restrictions because you have to keep
sure that the guest doesn't overwrite them accidentally.

Providing a copy of all ACPI tables by FwCfg isn't possible. ACPI tables
have to point to each other. So, if the guest copies the ACPI tables
into memory by it's own, it has to patch the tables. Due to different
layouts for different ACPI tables, there's no generic way to do that.
For that reason, qemu created a table loader interface. It contains
commands for the guest for loading specific blobs into guest memory and
patching those blobs.

This commit adds a qemu_loader class which handles the creation of qemu
loader commands. At the moment, the WRITE_POINTER command isn't
implement. It won't be required by bhyve's ACPI table generation yet.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>