bhyve/tpm: rename tpm2 to tpm
It was a bad idea to use a tpm2 suffix for all structs and files. The
tpm emulation for other tpm versions is similar. So, it's a better idea
to use a version variable in one of the tpm structs.

We don't want to support other versions than version 2.0 yet. So, the
change seems to be unnecessary. Nevertheless, we should use tpm instead
of tpm2 for upstreaming the changes and maybe we like to support newer
versions of the tpm in the future.

Note: For backward compatibility the command line and config node option
"tpm2" is still supported.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>