bhyve/tpm: emulate tpm passthru by accessing /dev/tpm0
The current tpm passthrough approach maps the physical mmio regions of
the physical tpm into the guest memory. This approach has some major
downsides. Bhyve has to implement an own emulation for every possible
interface. That's much work and hard to maintain.

This commit introduces another approach. Bhyve traps all TPM accesses
and always emulates the same tpm interface. To send commands to the tpm
device, bhyve uses the host tpm driver. So, only the host tpm driver has
to detect and interact with the physical tpm interface. The host tpm
driver has to deal with that regardless of bhyve which makes it a better
approach than the old one.

In order to keep this commit small and readable, this is just the first
step toward the new tpm emulation approach. Other commits will enable
the trap for tpm register and removes the mapping of the tpm mmio
regions.

Signed-off-by: Corvin Köhne <c.koehne@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>