x86/pci: add sysctl for intel_graphics_gen
Some register have changed for intel graphics of generation 11 and newer. Most
notably, the BDSM (Base Data Stolen Memory) register is located at a different
address and was changed from 32 bit to 64 bit. In order to properly emulate
this, bhyve has to know the generation of the intel graphics device. As we're
already filtering the devices for generation here, we can just add the
generation information and expose it as sysctl.

We know that the generation is not properly reported by the new sysctl. Some
generation don't differ much, so FreeBSD doesn't distinguish between them. As
we only need the sysctl for bhyve to check for generation 11 or newer, that's
totally fine for now. We don't want to take the hassle to split all device id's
without a use case. If a use case comes up where a more detailed value is
required, we can fix that easily and it shouldn't break old user.

Reviewed by:		<If someone else reviewed your modification.>
MFC after:		1 week
Sponsored by:		Beckhoff Automation GmbH & Co. KG
Differential Revision:	https://reviews.freebsd.org/D45333