bhyve: avoid resource leak in error path
In e820_finalize(), the e820_fwcfg_item variable, containing the etc/e820 file
(for the e820 table from the BIOS) is not free()'d when it could not be added
to the QEMU firmware configuration device (fw_cfg).

The only user of this function I could find is bhyve_init_platform_late(),
which is called by main() in bhyverun.c, which exits upon errors. (4 "exited
due to an error")

Therefore, as things are this fix looks superfluous, but it would quiet
Coverity down for this code path.

Reviewed by:		corvink, jhb
MFC after:		1 week
Differential Revision:	https://reviews.freebsd.org/D45223